requirements.txt
# Core LangChain packages
langchain-experimental
langchain-community
langchain-openai

# OpenAI SDK – must be >=1.40.0 for max_completion_tokens support
openai

# Azure + DB + ML dependencies
azure-functions
azure-search-documents==11.4.0
azure-identity
pyodbc==5.1.0
scikit-learn
sklearn-model==0.0.6

# Utilities
python-dotenv
deep-translator==1.11.4
langdetect==1.0.9
httpx==0.27.2


azure-ai-textanalytics==5.3.0
azure-ai-translation-text==1.0.1